pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "internship-finder"
version = "1.0.0"
description = "Underclass Internship Scanner - finds freshman/sophomore targeted internships"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "dateparser>=1.2.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "jinja2>=3.1.0",
    "pandas>=2.1.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "anthropic>=0.18.0",
    "pyyaml>=6.0.0",
    "sendgrid>=6.11.0",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
internship-scanner = "app.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
